pyproject.toml
[project]
name = "dataplatform"
version = "0.1.0"
description = "Algorithmic Trading Platform - Data & ML Infrastructure"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0", # YAML config loader
    # Database clients
    "clickhouse-driver>=0.2.6",
    "redis[hiredis]>=5.0.0",
    "psycopg2-binary>=2.9.9",
    # WebSocket
    "websockets>=12.0",
    # AWS SDK (async)
    "aioboto3>=12.0.0",
    "botocore>=1.31.0",
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "psutil>=5.9.0", # For integration tests (process management)
    "boto3>=1.34.0", # For Kinesis integration tests
    "aiokafka>=0.12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
